configs: bthomehubv5a: change configuration to boot from UBI partition
With the recent changes in LEDE the kernel is now stored inside a UBI
partition (volume). Enable and configure the MTD and UBI settings to
allow loading the kernel from the UBI volume with the name "kernel" and
change the default boot command to use the "kernel" UBI volume.
While here, also update the environment partition offset and size to
match the stock u-boot so we can re-use the same partitioning schema in
LEDE.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>